chore(FileDropZone): refactor, add qa, add snapshots
