+++ src/components/lab/FileDropZone/__stories__/constants.ts
@@ -1,0 +1,1 @@
+export const FileDropZoneQA = 'file-drop-zone-qa';
