+++ src/components/lab/FileDropZone/index.ts
@@ -1,1 +1,2 @@
 export * from './FileDropZone';
+export type {FileDropZoneProps, DropZoneFileRejection} from './types';
